+++ auth/src/main/AndroidManifest.xml
@@ -49,11 +49,6 @@
             android:label="@string/default_toolbar_title"
             android:theme="@style/FirebaseUI.Translucent" />
         <activity
-            android:name=".ui.ChooseAccountActivity"
-            android:label="@string/default_toolbar_title"
-            android:theme="@style/FirebaseUI.Translucent" />
-
-        <activity
             android:name="com.facebook.FacebookActivity"
             android:configChanges="keyboard|keyboardHidden|screenLayout|screenSize|orientation"
             android:label="@string/app_name"
