+++ auth/src/main/java/com/firebase/ui/auth/AuthUI.java
@@ -21,18 +21,17 @@ import android.support.annotation.DrawableRes;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.StyleRes;
-import android.support.annotation.VisibleForTesting;
 
 import com.facebook.FacebookSdk;
 import com.facebook.login.LoginManager;
 import com.firebase.ui.auth.provider.IDPProviderParcel;
-import com.firebase.ui.auth.ui.ChooseAccountActivity;
 import com.firebase.ui.auth.ui.FlowParameters;
 import com.firebase.ui.auth.ui.idp.AuthMethodPickerActivity;
 import com.firebase.ui.auth.util.CredentialsApiHelper;
 import com.firebase.ui.auth.util.GoogleApiClientTaskHelper;
 import com.firebase.ui.auth.util.Preconditions;
 import com.firebase.ui.auth.util.ProviderHelper;
+import com.firebase.ui.auth.util.SignInDelegate;
 import com.google.android.gms.auth.api.Auth;
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
 import com.google.android.gms.common.api.GoogleApiClient;
@@ -57,35 +56,35 @@ import java.util.Set;
  * If an alternative app instance is in use, call
  * {@link AuthUI#getInstance(FirebaseApp) AuthUI.getInstance(app} instead, passing the
  * appropriate app instance.
- *
+ * <p>
  * <h2>Sign-in</h2>
- *
+ * <p>
  * If a user is not currently signed in (as can be determined by checking
  * {@code auth.getCurrentUser() != null}, where {@code auth} is the {@link FirebaseAuth}
  * associated with your {@link FirebaseApp}) then the sign-in process can be started by creating
  * a sign-in intent using {@link SignInIntentBuilder}. A builder instance can be retrieved by
  * calling {@link AuthUI#createSignInIntentBuilder()}.
- *
+ * <p>
  * <p>The builder provides the following customization options for the authentication flow
  * implemented by this library:
- *
+ * <p>
  * <ul>
- *     <li>The set of authentication methods desired can be specified.</li>
- *     <li>The terms of service URL for your app can be specified, which is included as a link
- *         in the small-print of the account creation step for new users. If no terms of service
- *         URL is provided, the associated small-print is omitted.
- *     </li>
- *     <li>A custom theme can specified for the flow, which is applied to all the activities in
- *         the flow for consistent customization of colors and typography.
- *     </li>
+ * <li>The set of authentication methods desired can be specified.</li>
+ * <li>The terms of service URL for your app can be specified, which is included as a link
+ * in the small-print of the account creation step for new users. If no terms of service
+ * URL is provided, the associated small-print is omitted.
+ * </li>
+ * <li>A custom theme can specified for the flow, which is applied to all the activities in
+ * the flow for consistent customization of colors and typography.
+ * </li>
  * </ul>
- *
- *
+ * <p>
+ * <p>
  * <h3>Sign-in examples</h3>
- *
+ * <p>
  * If no customization is required, and only email authentication is required, the sign-in flow
  * can be started as follows:
- *
+ * <p>
  * <pre>
  * {@code
  * startActivityForResult(
@@ -93,9 +92,9 @@ import java.util.Set;
  *     RC_SIGN_IN);
  * }
  * </pre>
- *
+ * <p>
  * If Google Sign-in and Facebook Sign-in are also required, then this can be replaced with:
- *
+ * <p>
  * <pre>
  * {@code
  * startActivityForResult(
@@ -106,9 +105,9 @@ import java.util.Set;
  *     RC_SIGN_IN);
  * }
  * </pre>
- *
+ * <p>
  * Finally, if a terms of service URL and a custom theme are required:
- *
+ * <p>
  * <pre>
  * {@code
  * startActivityForResult(
@@ -121,15 +120,15 @@ import java.util.Set;
  *     RC_SIGN_IN);
  * }
  * </pre>
- *
+ * <p>
  * <h3>Handling the Sign-in response</h3>
- *
+ * <p>
  * The authentication flow provides only two response codes: {@link Activity#RESULT_OK RESULT_OK}
  * if a user is signed in, and {@link Activity#RESULT_CANCELED RESULT_CANCELLED} if sign in
  * failed. No further information on failure is provided as it is not typically useful; the only
  * recourse for most apps if sign in fails is to ask the user to sign in again later, or proceed
  * with an anonymous account if supported.
- *
+ * <p>
  * <pre>
  * {@code
  * @Override
@@ -148,29 +147,29 @@ import java.util.Set;
  *   }
  * }
  * </pre>
- *
+ * <p>
  * <h2>Sign-out</h2>
- *
+ * <p>
  * With the integrations provided by AuthUI, signing out a user is a multi-stage process:
- *
+ * <p>
  * <ol>
- *     <li>The user must be signed out of the {@link FirebaseAuth} instance.</li>
- *     <li>Smart Lock for Passwords must be instructed to disable automatic sign-in, in
- *         order to prevent an automatic sign-in loop that prevents the user from switching
- *         accounts.
- *     </li>
- *     <li>If the current user signed in using either Google or Facebook, the user must also be
- *         signed out using the associated API for that authentication method. This typically
- *         ensures that the user will not be automatically signed-in using the current account
- *         when using that authentication method again from the authentication method picker, which
- *         would also prevent the user from switching between accounts on the same provider.
- *     </li>
+ * <li>The user must be signed out of the {@link FirebaseAuth} instance.</li>
+ * <li>Smart Lock for Passwords must be instructed to disable automatic sign-in, in
+ * order to prevent an automatic sign-in loop that prevents the user from switching
+ * accounts.
+ * </li>
+ * <li>If the current user signed in using either Google or Facebook, the user must also be
+ * signed out using the associated API for that authentication method. This typically
+ * ensures that the user will not be automatically signed-in using the current account
+ * when using that authentication method again from the authentication method picker, which
+ * would also prevent the user from switching between accounts on the same provider.
+ * </li>
  * </ol>
- *
+ * <p>
  * In order to make this process easier, AuthUI provides a simple
  * {@link AuthUI#signOut(Activity) signOut} method to encapsulate this behavior. The method returns
  * a {@link Task} which is marked completed once all necessary sign-out operations are completed:
- *
+ * <p>
  * <pre>
  * {@code
  * public void onClick(View v) {
@@ -186,15 +185,15 @@ import java.util.Set;
  *   }
  * }
  * </pre>
- *
+ * <p>
  * <h2>IDP Provider configuration</h2>
- *
+ * <p>
  * Interacting with identity providers typically requires some additional client configuration.
  * AuthUI currently supports Google Sign-in and Facebook Sign-in, and currently requires the
  * basic configuration for these providers to be specified via string properties:
- *
+ * <p>
  * <ul>
- *
+ * <p>
  * <li>Google Sign-in: If your app build uses the
  * <a href="https://developers.google.com/android/guides/google-services-plugin">Google
  * Services Gradle Plugin</a>, no additional configuration is required. If not, please override
@@ -202,17 +201,20 @@ import java.util.Set;
  * <a href="https://developers.google.com/identity/sign-in/web/devconsole-project">Google OAuth
  * web client id.</a>
  * </li>
- *
+ * <p>
  * <li>Facebook Sign-in: Please override the string resource
  * {@code facebook_application_id} to provide the
  * <a href="https://developers.facebook.com/docs/apps/register">App ID</a> for your app as
  * registered on the
  * <a href="https://developers.facebook.com/apps">Facebook Developer Dashboard</a>.
  * </li>
- *
+ * <p>
  * </ul>
  */
 public class AuthUI {
+    public interface AuthUIResult {
+        void onResult(int resultCode, Intent data);
+    }
 
     /**
      * Provider identifier for email and password credentials, for use with
@@ -312,6 +314,7 @@ public class AuthUI {
     /**
      * Retrieves the {@link AuthUI} instance associated with the default app, as returned by
      * {@code FirebaseApp.getInstance()}.
+     *
      * @throws IllegalStateException if the default app is not initialized.
      */
     public static AuthUI getInstance() {
@@ -337,7 +340,8 @@ public class AuthUI {
      * Default theme used by {@link SignInIntentBuilder#setTheme(int)} if no theme
      * customization is required.
      */
-    public static @StyleRes int getDefaultTheme() {
+    @StyleRes
+    public static int getDefaultTheme() {
         // TODO(iainmgin): figure out why this works as a static method but not as a static
         //                 final variable.
         return R.style.FirebaseUI;
@@ -352,8 +356,6 @@ public class AuthUI {
         private List<String> mProviders = Collections.singletonList(EMAIL_PROVIDER);
         private String mTosUrl;
         private boolean mIsSmartLockEnabled = true;
-
-        private SignInIntentBuilder() {}
 
         /**
          * Specifies the theme to use for the application flow. If no theme is specified,
@@ -389,7 +391,7 @@ public class AuthUI {
          * Specifies the set of supported authentication providers. At least one provider
          * must be specified, and the set of providers must be a subset of
          * {@link #SUPPORTED_PROVIDERS}.
-         *
+         * <p>
          * <p>If no providers are explicitly specified by calling this method, then
          * {@link #EMAIL_PROVIDER email} is the default supported provider.
          *
@@ -409,7 +411,7 @@ public class AuthUI {
 
         /**
          * Enables or disables the use of Smart Lock for Passwords in the sign in flow.
-         * 
+         * <p>
          * <p>SmartLock is enabled by default
          */
         public SignInIntentBuilder setIsSmartLockEnabled(boolean enabled) {
@@ -417,24 +419,22 @@ public class AuthUI {
             return this;
         }
 
-        public Intent build() {
+        public void build(Activity activity, AuthUIResult result) {
             Context context = mApp.getApplicationContext();
-            return build(context);
+            build(context, activity, result);
         }
 
-        @VisibleForTesting
-        public Intent build(Context context) {
+        private void build(Context context, Activity activity, AuthUIResult result) {
             List<IDPProviderParcel> providerInfo =
                     ProviderHelper.getProviderParcels(context, mProviders);
-            return ChooseAccountActivity.createIntent(
-                    context,
-                    new FlowParameters(
-                            mApp.getName(),
-                            providerInfo,
-                            mTheme,
-                            mLogo,
-                            mTosUrl,
-                            mIsSmartLockEnabled));
+            SignInDelegate.newInstance(activity, result,
+                                       new FlowParameters(
+                                               mApp.getName(),
+                                               providerInfo,
+                                               mTheme,
+                                               mLogo,
+                                               mTosUrl,
+                                               mIsSmartLockEnabled));
         }
     }
 }
