--- auth/src/main/java/com/firebase/ui/auth/util/smartlock/RetrieveSmartLock.java
@@ -1,4 +1,0 @@
-package com.firebase.ui.auth.util.smartlock;
-
-public class RetrieveSmartLock {
-}
