+++ auth/src/test/java/com/firebase/ui/auth/ui/ChooseAccountActivityTest.java
@@ -25,7 +25,6 @@ import com.firebase.ui.auth.test_helpers.FakeAuthResult;
 import com.firebase.ui.auth.test_helpers.TestConstants;
 import com.firebase.ui.auth.test_helpers.TestHelper;
 import com.firebase.ui.auth.ui.idp.IDPSignInContainerActivity;
-import com.firebase.ui.auth.util.CredentialsAPI;
 import com.firebase.ui.auth.util.PlayServicesHelper;
 import com.google.android.gms.auth.api.credentials.IdentityProviders;
 import com.google.firebase.FirebaseApp;
