+++ auth/src/test/java/com/firebase/ui/auth/ui/email/SignInActivityTest.java
@@ -30,7 +30,6 @@ import com.firebase.ui.auth.test_helpers.TestConstants;
 import com.firebase.ui.auth.test_helpers.TestHelper;
 import com.firebase.ui.auth.ui.ExtraConstants;
 import com.firebase.ui.auth.util.PlayServicesHelper;
-import com.firebase.ui.auth.util.smartlock.SmartLock;
 import com.google.firebase.auth.FirebaseUser;
 
 import org.junit.Before;
